refactor(frontend): extract API + auth helpers to lib/api.js (1/N App.jsx split)
First step of the App.jsx monolith split (CLAUDE.md tech-debt #1, 3886 LOC
in a single file). This PR moves only the API base-URL resolution and
vault-auth helpers to a leaf module — pure JavaScript, zero React deps,
zero behavior change.

What moves to frontend/src/lib/api.js:
- DEFAULT_RENDER_URL  — fallback for SetupPanel default
- STATIC_URL           — GitHub Pages JSON fallback path
- RENDER_API           — resolved API base, env+localStorage chain
- apiToken / authHeaders — vault Bearer-token helpers

Same lookup order, same localStorage keys, same trailing-slash strip.
Existing `${RENDER_API}/api/...` template strings throughout App.jsx
work unchanged.

App.jsx change:
- 1 new import line
- ~14 LOC of helper definitions removed (replaced with comment crumbs
  pointing at lib/api.js so future readers find it)

Vite build: clean, 4.68s, identical bundle output (no functional change).

Roadmap for remaining App.jsx PRs:
  2.  lib/format.js   — fmtSal, timeAgo, salary helpers
  3.  lib/theme.js    — TH theme tokens
  4.  components/BrandLogo.jsx
  5.  components/SetupPanel.jsx
  6.  components/VaultRow.jsx
  7.  components/JobCard.jsx
  8-16. per-tab modules (Monitor, Pipeline, Trends, Analytics, Companies,
        Rare, Tracker, Vault, Jobs)
